# Hooks: Optimize addHook as appending operation