--- a/packages/hooks/benchmark/index.js
+++ b/packages/hooks/benchmark/index.js
@@ -3,17 +3,47 @@ const hooks = require( '../' );
 
 const suite = new Benchmark.Suite;
 
-function myCallback() {}
+const filter = process.argv[ 2 ];
+const isInFilter = ( key ) => ! filter || filter === key;
 
-hooks.addFilter( 'handled', 'myCallback', myCallback );
+function reset() {
+	hooks.removeAllFilters( 'example' );
+	hooks.removeAllActions( 'example' );
+}
+
+if ( isInFilter( 'applyFilters' ) ) {
+	function myCallback() {}
+
+	hooks.addFilter( 'example', 'myCallback', myCallback );
+
+	suite
+		.add( 'applyFilters - handled', () => {
+			hooks.applyFilters( 'handled' );
+		} )
+		.add( 'applyFilters - unhandled', () => {
+			hooks.applyFilters( 'unhandled' );
+		} );
+}
+
+if ( isInFilter( 'addFilter' ) ) {
+	let hasSetHighPriority = false;
+
+	suite
+		.add( 'addFilter - append last', () => {
+			hooks.addFilter( 'example', 'myCallback', () => {} );
+		} )
+		.add( 'addFilter - default before higher priority', () => {
+			if ( ! hasSetHighPriority ) {
+				hasSetHighPriority = true;
+				hooks.addFilter( 'example', 'priority', () => {}, 20 );
+			}
+
+			hooks.addFilter( 'example', 'myCallback', () => {} );
+		} );
+}
 
 suite
-	.add( 'handled', () => {
-		hooks.applyFilters( 'handled' );
-	} )
-	.add( 'unhandled', () => {
-		hooks.applyFilters( 'unhandled' );
-	} )
+	.on( 'cycle', reset )
 	// eslint-disable-next-line no-console
 	.on( 'cycle', ( event ) => console.log( event.target.toString() ) )
 	.run( { async: true } );