@gziolo 2018-12-13
> * Tracking handlers as an arrangement of `handlers[ priority ] = [ handlersForPriority ]`, leveraging the behavior of `Array#forEach` in sparse arrays

> However, these are technically breaking changes, since we had chosen to both expose and document `actions` and `filters` directly from the module. In retrospect, these should not have been exposed, as they are implementation details.

Yes, it would be much simpler if we could group them by priority. However, it might turn out that trying to keep backward compatibility might become less performant :(