@aduth 2018-12-13
> However, it might turn out that trying to keep backward compatibility might become less performant :(

There are possibly other options where we define the actions (or handlers specifically) on the module exports using `Object.defineProperty` such that the existing shape is respected but only computed on-demand.

I'd kinda see the internals supporting this anyways, where the linked list could be implemented via a generator, which can be converted to an array pretty easily.

```js
function* iterateLinkedList( head ) {
	let current = head;
	do {
		yield current.data;
		current = current.next;
	} while ( current );
}

const headHandler = { data: fn, next: { data: fn, next: null } };
const handlersAsArray = [ ...iterateHandlers( headHandler ) ]; // Or `Array.from`
// [ fn, fn ]
```