@gziolo 2018-12-13
Good point made, I didn't think about this. Indeed, we can expose a wrapper in those lines as described in your comment:
https://github.com/WordPress/gutenberg/blob/master/packages/hooks/src/createHooks.js#L37-L38

I'll leave it up to you, whether you want to tackle it in this PR or open a follow-up.